+++ sources/plugins/EnOcean/profiles/hardCoded/Profile_D2_01_04.cpp
@@ -89,11 +89,11 @@ void CProfile_D2_01_04::sendConfiguration(const shared::CDataContainer& deviceCo
                                           const std::string& senderId,
                                           boost::shared_ptr<IMessageHandler> messageHandler) const
 {
-   auto localControl = deviceConfiguration.get<std::string>("localControl") == "enable";
-   auto taughtInAllDevices = deviceConfiguration.get<std::string>("taughtIn") == "allDevices";
-   auto dimTimer1 = deviceConfiguration.get<double>("dimTimer1");
-   auto dimTimer2 = deviceConfiguration.get<double>("dimTimer2");
-   auto dimTimer3 = deviceConfiguration.get<double>("dimTimer3");
+   const auto localControl = deviceConfiguration.get<std::string>("localControl") == "enable";
+   const auto taughtInAllDevices = deviceConfiguration.get<std::string>("taughtIn") == "allDevices";
+   const auto dimTimer1 = deviceConfiguration.get<double>("dimTimer1");
+   const auto dimTimer2 = deviceConfiguration.get<double>("dimTimer2");
+   const auto dimTimer3 = deviceConfiguration.get<double>("dimTimer3");
 
    // CMD 0x2 - Actuator Set Local
    CProfile_D2_01_Common::sendActuatorSetLocalCommand(messageHandler,